Fixed issues with Perk parser and added new commands
* Added two new commands. !set_perk_cache_duration and !reset_perk_cache.
* Implemented workaround for the issue that prevents reading the content of perk log.
* Add new function to get relative time.
* Some feedback improvements

+++ src/Model.cs
@@ -11,7 +11,7 @@ namespace Settings
 
     public class ServerScheduleSettings
     {
-        public ulong ServerRestartSchedule = Convert.ToUInt64(TimeSpan.FromHours(6).TotalMinutes);
+        public uint ServerRestartSchedule = Convert.ToUInt32(TimeSpan.FromHours(6).TotalMinutes);
     }
 
     public class BotSettings
